32 bit -> 64 bit in values that top billions
lol oops
taking away presents lmao

+++ ChristmasClicker/ClickerSystem.cs
@@ -17,9 +17,9 @@ namespace ChristmasClicker
 
         public static int ticksBeforeClick;
 
-        public static int td;
-        public static int tdlag;
-        public static int totalDelivered
+        public static long td;
+        public static long tdlag;
+        public static long totalDelivered
         {
             get
             {
